+++ src/imio/helpers/tests/test_workflow.py
@@ -3,6 +3,7 @@ from imio.helpers.testing import IntegrationTestCase
 from imio.helpers.workflow import do_transitions
 from imio.helpers.workflow import get_state_infos
 from imio.helpers.workflow import get_transitions
+from imio.helpers.workflow import load_workflow_from_package
 from plone import api
 
 
@@ -42,3 +43,15 @@ class TestWorkflowModule(IntegrationTestCase):
         self.assertEqual(api.content.get_state(obj), 'external')
         do_transitions(obj, ['submit', 'retract', 'hide'])
         self.assertEqual(api.content.get_state(obj), 'private')
+
+    def test_load_workflow_from_package(self):
+        wkf_tool = api.portal.get_tool('portal_workflow')
+        wkf_obj = wkf_tool.get('intranet_workflow')
+        states = wkf_obj.states
+        self.assertIn('internal', states)
+        states.deleteStates(['internal'])
+        self.assertNotIn('internal', states)
+        self.assertTrue(load_workflow_from_package('intranet_workflow', 'profile-Products.CMFPlone:plone'))
+        wkf_obj = wkf_tool.get('intranet_workflow')
+        states = wkf_obj.states
+        self.assertIn('internal', states)